sage.misc.timing: Split out from sage.misc.misc
