+++ pkgs/sagemath-objects/MANIFEST.in
@@ -83,6 +83,7 @@ graft sage/libs/gmp
 
 ## FIXME: Needed for doctesting
 include sage/misc/misc.*              # walltime, cputime used in sage.doctest
+include sage/misc/timing.p*
 
 
 global-exclude *.c
