+++ src/sage/databases/cremona.py
@@ -51,7 +51,7 @@ from sage.misc.prandom import randint
 import sage.schemes.elliptic_curves.constructor as elliptic
 from .sql_db import SQLDatabase, verify_column
 from sage.features.databases import DatabaseCremona
-from sage.misc.misc import walltime
+from sage.misc.timing import walltime
 
 import re
 import string
