+++ src/sage/matrix/benchmark.py
@@ -21,7 +21,7 @@ from .constructor import random_matrix, Matrix
 from sage.rings.integer_ring import ZZ
 from sage.rings.rational_field import QQ
 from sage.rings.finite_rings.finite_field_constructor import FiniteField as GF
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 from cysignals.alarm import AlarmInterrupt, alarm, cancel_alarm
 
 from sage.interfaces.magma import magma
