+++ src/sage/matrix/matrix_integer_dense_hnf.py
@@ -10,7 +10,7 @@ from copy import copy
 
 from sage.arith.misc import CRT_list, previous_prime
 from sage.matrix.constructor import identity_matrix, matrix, random_matrix
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 from sage.misc.verbose import verbose
 from sage.rings.integer import Integer
 from sage.rings.integer_ring import ZZ
