+++ src/sage/misc/all__sagemath_objects.py
@@ -32,3 +32,5 @@ from .unknown import Unknown, UnknownError
 from .cachefunc import CachedFunction, cached_function, cached_method, cached_in_parent_method, disk_cached_function
 
 from .abstract_method import abstract_method
+
+from .timing import walltime, cputime
