+++ src/sage/misc/profiler.py
@@ -15,7 +15,7 @@ AUTHORS:
 #                  http://www.gnu.org/licenses/
 #*****************************************************************************
 
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 import inspect
 import sys
