+++ src/sage/modular/arithgroup/tests.py
@@ -19,7 +19,7 @@ from sage.modular.arithgroup.all import Gamma, Gamma0, Gamma1, GammaH
 from sage.rings.finite_rings.integer_mod_ring import Zmod
 
 import sage.misc.prandom as prandom
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 
 
 def random_even_arithgroup(index, nu2_max=None, nu3_max=None):
