+++ src/sage/modular/overconvergent/hecke_series.py
@@ -74,7 +74,7 @@ from sage.functions.other import floor, ceil
 from sage.matrix.constructor import matrix, random_matrix
 from sage.matrix.matrix_space import MatrixSpace
 from sage.misc.functional import dimension, transpose, charpoly
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 from sage.misc.verbose import verbose
 from sage.modular.dims import dimension_modular_forms
 from sage.modular.modform.all import ModularForms, ModularFormsRing, delta_qexp, eisenstein_series_qexp
