+++ src/sage/modular/quatalg/brandt.py
@@ -1690,7 +1690,7 @@ def benchmark_sage(levels, silent=False):
         ('sage', 43, 2, ...)
         ('sage', 97, 2, ...)
     """
-    from sage.misc.misc import cputime
+    from sage.misc.timing import cputime
     ans = []
     for p, M in levels:
         t = cputime()
