+++ src/sage/rings/padics/pow_computer_ext.pyx
@@ -60,7 +60,7 @@ from cysignals.signals cimport sig_on, sig_off
 
 include "sage/libs/ntl/decl.pxi"
 
-from sage.misc.misc import cputime
+from sage.misc.timing import cputime
 from sage.libs.gmp.mpz cimport *
 from sage.libs.ntl.ntl_ZZ_pContext cimport ntl_ZZ_pContext_factory
 from sage.libs.ntl.ntl_ZZ_pContext import ZZ_pContext_factory
